create new .js files and divide into moduls
