+++ js/util.js
@@ -1,0 +1,19 @@
+const getRandomNumber = (min, max) => {
+  if (min >= max) {
+    [min, max] = [max, min];
+  }
+  return Math.floor(Math.random() * (max - min + 1)) + min;
+}
+
+export {getRandomNumber};
+
+const getRandomItem = (array) => {
+    let randomIndex =getRandomNumber(0, array.length);
+    return array[randomIndex];
+}
+
+export {getRandomItem};
+
+const maxSymbols = (str, maxLength) => {
+  return str.length <= maxLength;
+}
